requirements.txt
# Generated on: Python 3.10.10
asn1crypto==1.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.2
filelock==3.10.0
idna==3.4
oscrypto==1.3.0
packaging==23.0
pycparser==2.21
pycryptodomex==3.17
PyJWT==2.6.0
pyOpenSSL==23.0.0
pytz==2022.7.1
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.15
snowflake-connector-python==3.0.2
yfinance==0.2.14
pandas-datareader==0.10.0
matplotlib==3.7.1
seaborn==0.12.2
lazypredict==0.2.12
plotly==5.14.0
streamlit-echarts==0.4.0
streamlit-option-menu==0.3.2
streamlit-lottie==0.0.3
tensorflow==2.12.0
yfinance==0.2.14